+++ src/api/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './auth';
 export * from './user';
 export * from './post';
+export * from './upload';
