+++ src/api/interfaces/upload.ts
@@ -1,0 +1,6 @@
+export interface ImgUpload {
+  method: 'POST';
+  endpoint: '/upload';
+  req: FormData;
+  res: {};
+};
