+++ src/api/serverResource.ts
@@ -1,4 +1,5 @@
 import {
+  Signup,
   LoginWithInfo,
   LoginWithRefreshToken,
   MyInfo,
@@ -7,9 +8,11 @@ import {
   UserFind,
   QuestionList,
   QuestionAnswer,
+  ImgUpload,
 } from './interfaces';
 
 export interface APIResource {
+  signup: Signup,
   loginWithInfo: LoginWithInfo;
   loginWithRefreshToken: LoginWithRefreshToken;
   myInfo: MyInfo;
@@ -18,4 +21,5 @@ export interface APIResource {
   userFind: UserFind;
   questionList: QuestionList;
   questionAnswer: QuestionAnswer;
+  imgUpload: ImgUpload;
 };
