+++ src/funtions/index.ts
@@ -1,3 +1,4 @@
 export * from './auth';
 export * from './exceptionPage';
 export { default as makeAlert } from './makeAlert';
+export * from './checkAuth';
