+++ src/main.tsx
@@ -11,6 +11,7 @@ import {
   Login,
   Components,
   Signup,
+  SignupImg,
   User,
 } from '@/pages';
 import { Screen, ExceptionPage } from '@/funtions';
@@ -50,6 +51,7 @@ const Router = () => {
         <Route path='/' element={<Screen Children={Main} />} />
         <Route path='/login' element={<ExceptionPage Children={Login} />} />
         <Route path='/signup' element={<ExceptionPage Children={Signup} />} />
+        <Route path='/imgUpload' element={<ExceptionPage Children={SignupImg} />} />
         <Route path='/components' element={<Components />} />
         <Route path='/:username' element={<Screen Children={User} />} />
       </Routes>
